+++ NEWS.md
@@ -4,6 +4,8 @@
 
 * Renamed items in effort list output of `das_effort` and `das_effort_sight`: 'siteinfo' is now 'sightinfo', and segdata columns indicating the number of sighitngs or animals on a segment are now formatted as 'description'_'species code', e.g. "nSI_075"
 
+* Fixed a bug in `das_sight` when processing "G" and "g" events
+
 
 # swfscDAS 0.1